+++ src/GSvar/SequencingRunWidget.cpp
@@ -12,13 +12,18 @@
 #include "GlobalServiceProvider.h"
 #include "BasicStatistics.h"
 #include "GSvarHelper.h"
+#include "NsxSettingsDialog.h"
 #include <numeric>
 
-SequencingRunWidget::SequencingRunWidget(QWidget* parent, QString run_id)
+SequencingRunWidget::SequencingRunWidget(QWidget* parent, const QStringList& run_ids)
 	: QWidget(parent)
 	, ui_(new Ui::SequencingRunWidget)
-	, run_id_(run_id)
+	, run_ids_(run_ids)
 {
+	if (run_ids_.size() < 1) THROW(ArgumentException, "At least one run_id has to be provided!");
+	is_batch_view_ = run_ids.size() > 1;
+
+	run_ids_.sort();
 	ui_->setupUi(this);
 	GUIHelper::styleSplitter(ui_->splitter);
 	ui_->splitter->setSizes(QList<int>() << 200 << 800);
@@ -46,12 +51,118 @@ SequencingRunWidget::SequencingRunWidget(QWidget* parent, QString run_id)
 	ui_->samples->addAction(action);
 	connect(action, SIGNAL(triggered(bool)), this, SLOT(scheduleForResequencing()));
 
+	if (is_batch_view_) initBatchView();
+
 	updateGUI();
+	ui_->gb_sequencing_run->setStyleSheet("QGroupBox {background-color: white;}");
 }
 
 SequencingRunWidget::~SequencingRunWidget()
 {
 	delete ui_;
+}
+
+void SequencingRunWidget::initBatchView()
+{
+	if (!is_batch_view_) return;
+	try
+	{
+		//clear default view
+		QLayoutItem* item;
+		while (ui_->gb_sequencing_run->layout()->count() > 0)
+		{
+			item = ui_->gb_sequencing_run->layout()->takeAt(0);
+			delete item->widget();
+			delete item;
+		}
+		delete ui_->name;
+		delete ui_->quality;
+		delete ui_->gb_sequencing_run->layout();
+
+		//replace sequencing run list with tablular view
+		QGridLayout* seq_run_table = new QGridLayout();
+
+		//create vertical header
+		int r = 0;
+		seq_run_table->addWidget(new QLabel("name:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("flowcell ID:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("flowcell type:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("device:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("side:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("recipe:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("start date:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("end date:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("pool molarity:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("status:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addWidget(new QLabel("comments:"), r++, 0, Qt::AlignLeft|Qt::AlignTop);
+		seq_run_table->addWidget(new QLabel("backup done:"), r++, 0, Qt::AlignLeft);
+		seq_run_table->addItem(new QSpacerItem(10, 20, QSizePolicy::Minimum, QSizePolicy::Expanding), r++, 0);
+
+
+		for (int c = 1; c <= run_ids_.size(); ++c)
+		{
+			r = 0; //reset row counter
+			QHBoxLayout* hbox = new QHBoxLayout();
+			QLabel* l_quality = new QLabel();
+			ProcessedSampleWidget::styleQualityLabel(l_quality, "n/a");
+			l_quality->setMaximumWidth(19);
+			hbox->addWidget(l_quality, 0, Qt::AlignLeft| Qt::AlignCenter);
+			QLabel* l_name = new QLabel("name");
+			l_name->setAlignment(Qt::AlignLeft| Qt::AlignCenter);
+			hbox->addWidget(l_name, 0, Qt::AlignLeft| Qt::AlignCenter);
+
+			QToolButton* btn_edit = new QToolButton();
+			btn_edit->setText("");
+			btn_edit->setIcon(QIcon(":/Icons/Edit.png"));
+			btn_edit->setIconSize(QSize(14,14));
+			btn_edit->setFixedWidth(19);
+			btn_edit->setFixedHeight(19);
+
+			//link button through signal mapper
+			QSignalMapper* signal_mapper = new QSignalMapper(this);
+			int run_id = run_ids_.at(c-1).toInt();
+			signal_mapper->setMapping(btn_edit, run_id);
+			connect(btn_edit, SIGNAL(clicked(bool)), signal_mapper, SLOT(map()));
+			connect(signal_mapper, SIGNAL(mapped(int)), this, SLOT(edit(int)));
+
+			hbox->addWidget(btn_edit, 0, Qt::AlignLeft| Qt::AlignCenter);
+
+			seq_run_table->addItem(hbox, r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("fcid"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("flowcell_type"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("d_name"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("side"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("recipe"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("start_date"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("end_date"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("molarity"), r++, c, Qt::AlignLeft);
+			seq_run_table->addWidget(new QLabel("status"), r++, c, Qt::AlignLeft);
+			QLabel* comments = new QLabel("comments");
+			comments->setAlignment(Qt::AlignLeft|Qt::AlignTop);
+			comments->setMaximumWidth(200);
+			seq_run_table->addWidget(comments, r++, c, Qt::AlignLeft|Qt::AlignTop);
+			seq_run_table->addWidget(new QLabel("backup_done"), r++, c, Qt::AlignLeft);
+			seq_run_table->setColumnMinimumWidth(c, 150);
+
+		}
+		seq_run_table->addItem(new QSpacerItem(20, 10, QSizePolicy::Expanding, QSizePolicy::Minimum), 0, run_ids_.size()+1);
+
+
+		ui_->gb_sequencing_run->setLayout(seq_run_table);
+		ui_->gb_sequencing_run->setStyleSheet("QGroupBox {background-color: white;}");
+		ui_->gb_sequencing_run->setMinimumHeight(280);
+
+
+
+		//disable edit/MID buttons in batch view
+		ui_->edit_btn->setEnabled(false);
+		ui_->mid_check_btn->setEnabled(false);
+
+	}
+	catch (Exception& e)
+	{
+		QMessageBox::warning(this, "Batch view init failed!", "Could not initialize widget:\n" + e.message());
+	}
 }
 
 void SequencingRunWidget::updateGUI()
@@ -61,39 +172,83 @@ void SequencingRunWidget::updateGUI()
 		//#### run details ####
 		NGSD db;
 		SqlQuery query = db.getQuery();
-		query.exec("SELECT r.*, d.name d_name, d.type d_type FROM sequencing_run r, device d WHERE r.device_id=d.id AND r.id='" + run_id_ + "'");
-		query.next();
-		ui_->name->setText(query.value("name").toString());
-		ui_->fcid->setText(query.value("fcid").toString());
-		ui_->fc_type->setText(query.value("flowcell_type").toString());
-		ui_->start->setText(query.value("start_date").toString());
-		ui_->end->setText(query.value("end_date").toString());
-		ui_->recipe->setText(query.value("recipe").toString());
-		GSvarHelper::limitLines(ui_->comments, query.value("comment").toString());
-		ui_->device->setText(query.value("d_name").toString() + " (" + query.value("d_type").toString() + ")");
-		ui_->side->setText(query.value("side").toString());
-		QVariant molarity = query.value("pool_molarity");
-		if (!molarity.isNull())
+		if (is_batch_view_)
 		{
-			ui_->molarity_and_method->setText(molarity.toString() + " (" + query.value("pool_quantification_method").toString() + ")");
+
+			//update data
+			int r;
+			for (int c = 1; c <= run_ids_.size(); ++c)
+			{
+				r = 0; //reset row counter
+				query.exec("SELECT r.*, d.name d_name, d.type d_type FROM sequencing_run r, device d WHERE r.device_id=d.id AND r.id='" + run_ids_.at(c-1) + "'");
+				query.next();
+
+				QGridLayout* seq_run_table = (QGridLayout*) ui_->gb_sequencing_run->layout();
+
+				QHBoxLayout* hbox = (QHBoxLayout*) seq_run_table->itemAtPosition(r++, c);
+				QLabel* l_quality = (QLabel*) hbox->itemAt(0)->widget();
+				ProcessedSampleWidget::styleQualityLabel(l_quality, query.value("quality").toString());
+				QLabel* l_name = (QLabel*) hbox->itemAt(1)->widget();
+				l_name->setText(query.value("name").toString());
+
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("fcid").toString());
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("flowcell_type").toString());
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("d_name").toString() + " (" + query.value("d_type").toString() + ")");
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("side").toString());
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("recipe").toString());
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("start_date").toString());
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("end_date").toString());
+				QVariant molarity = query.value("pool_molarity");
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText((molarity.isNull())?(molarity.toString() + " (" + query.value("pool_quantification_method").toString() + ")"):(""));
+				if (query.value("status").toString()=="analysis_finished")((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText(query.value("status").toString());
+				else ((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText("<font color=orange>" + query.value("status").toString() + "</font>");
+				GSvarHelper::limitLines((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget(), query.value("comment").toString());
+				((QLabel*) seq_run_table->itemAtPosition(r++, c)->widget())->setText((query.value("backup_done").toString()=="1" ? "yes" : "<font color=red>no</font>"));
+
+			}
+
+			//#### deactivate SampleSheet ####
+			ui_->novaseqx_samplesheet_btn->setEnabled(false);
+
 		}
 		else
 		{
-			ui_->molarity_and_method->setText("");
-		}
-		ProcessedSampleWidget::styleQualityLabel(ui_->quality, query.value("quality").toString());
-		QString status = query.value("status").toString();
-		ui_->status->setText(status);
-		ui_->backup->setText(query.value("backup_done").toString()=="1" ? "yes" : "no");
+			query.exec("SELECT r.*, d.name d_name, d.type d_type FROM sequencing_run r, device d WHERE r.device_id=d.id AND r.id='" + run_ids_.at(0) + "'");
+			query.next();
 
-		//#### activate SampleSheet ####
-		ui_->novaseqx_samplesheet_btn->setEnabled((query.value("d_type").toString() == "NovaSeqXPlus") || (query.value("d_type").toString() == "NovaSeqX"));
+			ui_->name->setText(query.value("name").toString());
+			ui_->fcid->setText(query.value("fcid").toString());
+			ui_->fc_type->setText(query.value("flowcell_type").toString());
+			ui_->start->setText(query.value("start_date").toString());
+			ui_->end->setText(query.value("end_date").toString());
+			ui_->recipe->setText(query.value("recipe").toString());
+			GSvarHelper::limitLines(ui_->comments, query.value("comment").toString());
+			ui_->device->setText(query.value("d_name").toString() + " (" + query.value("d_type").toString() + ")");
+			ui_->side->setText(query.value("side").toString());
+			QVariant molarity = query.value("pool_molarity");
+			if (!molarity.isNull())
+			{
+				ui_->molarity_and_method->setText(molarity.toString() + " (" + query.value("pool_quantification_method").toString() + ")");
+			}
+			else
+			{
+				ui_->molarity_and_method->setText("");
+			}
+			ProcessedSampleWidget::styleQualityLabel(ui_->quality, query.value("quality").toString());
+			QString status = query.value("status").toString();
+			ui_->status->setText(status);
+			ui_->backup->setText(query.value("backup_done").toString()=="1" ? "yes" : "no");
+
+			//#### activate SampleSheet ####
+			ui_->novaseqx_samplesheet_btn->setEnabled((query.value("d_type").toString() == "NovaSeqXPlus") || (query.value("d_type").toString() == "NovaSeqX"));
+		}
 
 		//#### run quality ####
 		updateReadQualityTable();
 
 		//#### sample table ####
 		updateRunSampleTable();
+
 	}
 	catch (Exception& e)
 	{
@@ -106,6 +261,7 @@ void SequencingRunWidget::updateRunSampleTable()
 {
 	//get data from NGSD
 	QStringList headers;
+	if (is_batch_view_) headers << "run";
 	headers << "lane" << "quality" << "sample" << "name external" << "is_tumor" << "is_ffpe" << "sample type" << "project" << "MID i7" << "MID i5" << "species" << "processing system" << "input [ng]" << "molarity [nM]" << "operator" << "processing modus" << "batch number" << "comments";
 	if (ui_->show_sample_comment->isChecked())
 	{
@@ -114,10 +270,27 @@ void SequencingRunWidget::updateRunSampleTable()
 	headers << "urgent";
 	headers << "resequencing";
 	NGSD db;
-	DBTable samples = db.createTable("processed_sample", "SELECT ps.id, ps.lane, ps.quality, CONCAT(s.name,'_',LPAD(ps.process_id,2,'0')), s.name_external, s.tumor, s.ffpe, s.gender, s.sample_type, (SELECT CONCAT(name, ' (', type, ')') FROM project WHERE id=ps.project_id), (SELECT CONCAT(name, ' (', sequence, ')') FROM mid WHERE id=ps.mid1_i7), (SELECT CONCAT(name, ' (', sequence, ')') FROM mid WHERE id=ps.mid2_i5), sp.name, sys.name_manufacturer, sys.type as sys_type, ps.processing_input, ps.molarity, (SELECT name FROM user WHERE id=ps.operator_id), ps.processing_modus, ps.batch_number, ps.comment" + QString(ui_->show_sample_comment->isChecked() ? ", s.comment as sample_comment" : "") + " ,ps.urgent, ps.scheduled_for_resequencing "+
-														  " FROM processed_sample ps, sample s, processing_system sys, species sp WHERE sp.id=s.species_id AND ps.processing_system_id=sys.id AND ps.sample_id=s.id AND ps.sequencing_run_id='" + run_id_ + "' "
-														  " ORDER BY ps.lane ASC, "+ (ui_->sort_by_ps_id->isChecked() ? "ps.id" : "ps.processing_system_id ASC, s.name ASC, ps.process_id"));
-	//format columns
+	DBTable samples;
+	if (is_batch_view_)
+	{
+		samples = db.createTable("processed_sample",  QString("SELECT ps.id, (SELECT name FROM sequencing_run WHERE id=ps.sequencing_run_id), ps.lane, ps.quality, ")
+														+ "CONCAT(s.name,'_',LPAD(ps.process_id,2,'0')), s.name_external, s.tumor, s.ffpe, s.gender, s.sample_type, "
+														+ "(SELECT CONCAT(name, ' (', type, ')') FROM project WHERE id=ps.project_id), "
+														+ "(SELECT CONCAT(name, ' (', sequence, ')') FROM mid WHERE id=ps.mid1_i7), (SELECT CONCAT(name, ' (', sequence, ')') FROM mid WHERE id=ps.mid2_i5), "
+														+ "sp.name, sys.name_manufacturer, sys.type as sys_type, ps.processing_input, ps.molarity, (SELECT name FROM user WHERE id=ps.operator_id), "
+														+ "ps.processing_modus, ps.batch_number, ps.comment" + QString(ui_->show_sample_comment->isChecked() ? ", s.comment as sample_comment" : "")
+														+ ",ps.urgent ,ps.scheduled_for_resequencing FROM processed_sample ps, sample s, processing_system sys, species sp "
+														+ "WHERE sp.id=s.species_id AND ps.processing_system_id=sys.id AND ps.sample_id=s.id AND ps.sequencing_run_id IN ('" + run_ids_.join("', '") + "') "
+														+ "ORDER BY ps.lane ASC, "+ (ui_->sort_by_ps_id->isChecked() ? "ps.id" : "ps.sequencing_run_id ASC, ps.processing_system_id ASC, s.name ASC, ps.process_id"));
+	}
+	else
+	{
+		samples = db.createTable("processed_sample", QString("SELECT ps.id, ps.lane, ps.quality, CONCAT(s.name,'_',LPAD(ps.process_id,2,'0')), s.name_external, s.tumor, s.ffpe, s.gender, s.sample_type, (SELECT CONCAT(name, ' (', type, ')') ")
+													+ "FROM project WHERE id=ps.project_id), (SELECT CONCAT(name, ' (', sequence, ')') FROM mid WHERE id=ps.mid1_i7), (SELECT CONCAT(name, ' (', sequence, ')') "
+													+ "FROM mid WHERE id=ps.mid2_i5), sp.name, sys.name_manufacturer, sys.type as sys_type, ps.processing_input, ps.molarity, (SELECT name FROM user WHERE id=ps.operator_id), ps.processing_modus, ps.batch_number, ps.comment" + QString(ui_->show_sample_comment->isChecked() ? ", s.comment as sample_comment" : "") + ", ps.urgent, ps.scheduled_for_resequencing "
+													+ "FROM processed_sample ps, sample s, processing_system sys, species sp WHERE sp.id=s.species_id AND ps.processing_system_id=sys.id AND ps.sample_id=s.id AND ps.sequencing_run_id IN ('" + run_ids_.join("', '") + "') "
+													+ "ORDER BY ps.lane ASC, "+ (ui_->sort_by_ps_id->isChecked() ? "ps.id" : "ps.processing_system_id ASC, s.name ASC, ps.process_id"));
+	}	//format columns
 	samples.formatBooleanColumn(samples.columnIndex("tumor"));
 	samples.formatBooleanColumn(samples.columnIndex("ffpe"));
 	samples.formatBooleanColumn(samples.columnIndex("urgent"), true);
@@ -125,7 +298,8 @@ void SequencingRunWidget::updateRunSampleTable()
 
 	// determine QC parameter based on sample types
 	QSet<QString> sample_types = samples.extractColumn(samples.columnIndex("sample_type")).toSet();
-	setQCMetricAccessions(sample_types);
+	QSet<QString> system_types = samples.extractColumn(samples.columnIndex("sys_type")).toSet();
+	setQCMetricAccessions(sample_types, system_types);
 
 	// update QC plot button
 	ui_->plot_btn->setMenu(new QMenu());
@@ -237,8 +411,9 @@ void SequencingRunWidget::updateRunSampleTable()
 	if (ui_->show_lab_cols->isChecked()) ui_->samples->setBackgroundColorIfEqual("resequencing", red, "yes");
 
 	//#### sample summary ####
-	QStringList imported_qc = db.getValues("SELECT ps.id FROM processed_sample ps WHERE ps.sequencing_run_id='" + run_id_ + "' AND EXISTS(SELECT id FROM processed_sample_qc WHERE processed_sample_id=ps.id)");
-	QStringList imported_vars = db.getValues("SELECT ps.id FROM processed_sample ps WHERE ps.sequencing_run_id='" + run_id_ + "' AND EXISTS(SELECT variant_id FROM detected_variant WHERE processed_sample_id=ps.id)");
+	QStringList imported_qc, imported_vars;
+	imported_qc = db.getValues("SELECT ps.id FROM processed_sample ps WHERE ps.sequencing_run_id IN ('" + run_ids_.join("', '") + "') AND EXISTS(SELECT id FROM processed_sample_qc WHERE processed_sample_id=ps.id)");
+	imported_vars = db.getValues("SELECT ps.id FROM processed_sample ps WHERE ps.sequencing_run_id IN ('" + run_ids_.join("', '") + "') AND EXISTS(SELECT variant_id FROM detected_variant WHERE processed_sample_id=ps.id)");
 	ui_->sample_count->setText(QString::number(samples.rowCount()) + " samples (" + QString::number(imported_qc.count()) + " with QC, " + QString::number(imported_vars.count()) + " with variants)");
 }
 
@@ -304,7 +479,14 @@ void SequencingRunWidget::showPlot()
 	}
 	if (selected_rows.isEmpty()) //no selection > select all samples (works if all have the sample processing system)
 	{
-		selected_ps_ids << db.getValues("SELECT id FROM processed_sample WHERE sequencing_run_id='" + run_id_ + "'");
+		if (is_batch_view_)
+		{
+			selected_ps_ids << db.getValues("SELECT id FROM processed_sample WHERE sequencing_run_id IN ('" + run_ids_.join("', '") + "')");
+		}
+		else
+		{
+			selected_ps_ids << db.getValues("SELECT id FROM processed_sample WHERE sequencing_run_id='" + run_ids_.at(0) + "'");
+		}
 	}
 
 	//check one processing system is selected
@@ -337,8 +519,27 @@ void SequencingRunWidget::showPlot()
 
 void SequencingRunWidget::edit()
 {
-	DBEditor* widget = new DBEditor(this, "sequencing_run", run_id_.toInt());
+	//not working in batch view
+	if (is_batch_view_) return;
+
+	DBEditor* widget = new DBEditor(this, "sequencing_run", run_ids_.at(0).toInt());
 	auto dlg = GUIHelper::createDialog(widget, "Edit sequencing run " + ui_->name->text() ,"", true);
+	if (dlg->exec()==QDialog::Accepted)
+	{
+		widget->store();
+		updateGUI();
+	}
+}
+
+void SequencingRunWidget::edit(int run_id)
+{
+	//only working in batch view
+	if (!is_batch_view_) return;
+
+	DBEditor* widget = new DBEditor(this, "sequencing_run", run_id);
+	QString run_name = NGSD().getValue("SELECT name FROM sequencing_run WHERE id=:0;", false, QString::number(run_id)).toString();
+
+	auto dlg = GUIHelper::createDialog(widget, "Edit sequencing run " + run_name ,"", true);
 	if (dlg->exec()==QDialog::Accepted)
 	{
 		widget->store();
@@ -348,32 +549,56 @@ void SequencingRunWidget::edit()
 
 void SequencingRunWidget::sendStatusEmail()
 {
-	QString status = ui_->status->text();
-	if (status!="analysis_finished" && status!="run_finished")
-	{
-		QMessageBox::information(this, "Run status email", "Run status emails can only be generated for runs with status 'analysis_finished' or 'run_finished'!");
-		return;
-	}
-
 	NGSD db;
-	QString run_name = ui_->name->text();
 
 	//create email
 	QStringList to, body;
 	QString subject;
+	QString status;
+	QString run_name;
 
-	if (status=="analysis_finished")
+	if (is_batch_view_)
+	{
+		// get run status of all runs
+		QSet<QString> run_statuses = db.getValues("SELECT status FROM sequencing_run WHERE id IN (" + run_ids_.join(", ") + ");").toSet();
+		run_statuses.remove("analysis_finished");
+		run_statuses.remove("run_finished");
+
+		if(run_statuses.size() > 0 )
+		{
+			QMessageBox::information(this, "Run status email", "Run status emails can only be generated if all runs have status 'analysis_finished' or 'run_finished'!");
+			return;
+		}
+
+		QStringList run_names = db.getValues("SELECT name FROM sequencing_run WHERE id IN (" + run_ids_.join(", ") + ");");
+		std::sort(run_names.begin(), run_names.end());
+		run_name = run_names.first() + " - " + run_names.last();
+
+	}
+	else
+	{
+		status = ui_->status->text();
+		if (status!="analysis_finished" && status!="run_finished")
+		{
+			QMessageBox::information(this, "Run status email", "Run status emails can only be generated for runs with status 'analysis_finished' or 'run_finished'!");
+			return;
+		}
+
+		run_name = ui_->name->text();
+	}
+
+	if (is_batch_view_ || status=="analysis_finished")
 	{
 		//update table info to be sure it is the current info.
 		updateRunSampleTable();
 
 		to << Settings::string("email_run_analyzed").split(";");
 
-		subject = "[NGSD] Lauf " + run_name + " analysiert";
+		subject = (is_batch_view_)?"[NGSD] Läufe " + run_name + " analysiert": "[NGSD] Lauf " + run_name + " analysiert";
 
 		body << "Hallo zusammen,";
 		body << "";
-		body << "der Lauf " + run_name + " ist fertig analysiert:";
+		body << ((is_batch_view_)?"die Läufe " + run_name + " sind fertig analysiert:": "der Lauf " + run_name + " ist fertig analysiert:");
 
 		body << "";
 
@@ -414,7 +639,7 @@ void SequencingRunWidget::sendStatusEmail()
 
 
 		SqlQuery query = db.getQuery();
-		query.exec("SELECT DISTINCT p.id, p.name, p.email_notification, p.internal_coordinator_id, p.analysis FROM project p, processed_sample ps WHERE ps.project_id=p.id AND ps.sequencing_run_id=" + run_id_ + " ORDER BY p.name ASC");
+		query.exec("SELECT DISTINCT p.id, p.name, p.email_notification, p.internal_coordinator_id, p.analysis FROM project p, processed_sample ps WHERE ps.project_id=p.id AND ps.sequencing_run_id IN ('" + run_ids_.join("', '") + "') ORDER BY p.name ASC");
 		while(query.next())
 		{
 			int coordinator_id = query.value("internal_coordinator_id").toInt();
@@ -425,11 +650,11 @@ void SequencingRunWidget::sendStatusEmail()
 			body << "Projekt: " + query.value("name").toString();
 			body << "  Koordinator: " + db.userName(coordinator_id);
 
-			int ps_count = db.getValue("SELECT count(id) FROM processed_sample WHERE sequencing_run_id='" + run_id_ + "' AND project_id='" + query.value("id").toString() +"'").toInt();
+			int ps_count = db.getValue("SELECT count(id) FROM processed_sample WHERE sequencing_run_id IN ('" + run_ids_.join("', '") + "') AND project_id='" + query.value("id").toString() +"'").toInt();
 			body << "  Proben: " + QString::number(ps_count);
 			body << "  Analyse: " + query.value("analysis").toString();
 
-			QStringList operator_ids = db.getValues("SELECT operator_id FROM processed_sample WHERE sequencing_run_id='" + run_id_ + "' AND project_id='" + query.value("id").toString() + "' AND operator_id IS NOT NULL");
+			QStringList operator_ids = db.getValues("SELECT operator_id FROM processed_sample WHERE sequencing_run_id IN ('" + run_ids_.join("', '") + "') AND project_id='" + query.value("id").toString() + "' AND operator_id IS NOT NULL");
 			operator_ids.removeDuplicates();
 			operator_ids.removeAll("");
 			foreach(QString operator_id, operator_ids)
@@ -456,20 +681,26 @@ void SequencingRunWidget::sendStatusEmail()
 	//send
 	EmailDialog dlg(this, to, subject, body);
 	dlg.exec();
+
 }
 
 void SequencingRunWidget::checkMids()
 {
+	if (is_batch_view_)
+	{
+		GUIHelper::showMessage("check MIDs", "MID check is not supported in batch view!");
+		return;
+	}
 	try
 	{
 		NGSD db;
 
 		//create dialog
 		MidCheckWidget* widget = new MidCheckWidget();
-		widget->addRun(db.getValue("SELECT name FROM sequencing_run WHERE id=" + run_id_).toString());
+		widget->addRun(db.getValue("SELECT name FROM sequencing_run WHERE id=" + run_ids_.at(0)).toString());
 
 		//determine usable length
-		QString recipe = db.getValue("SELECT recipe FROM sequencing_run WHERE id=" + run_id_).toString();
+		QString recipe = db.getValue("SELECT recipe FROM sequencing_run WHERE id=" + run_ids_.at(0)).toString();
 		QPair<int,int> lengths = MidCheck::lengthFromRecipe(recipe);
 		QPair<int,int> tmp = MidCheck::lengthFromSamples(widget->mids());
 		lengths.first = std::min(lengths.first, tmp.first);
@@ -488,15 +719,29 @@ void SequencingRunWidget::checkMids()
 
 void SequencingRunWidget::exportSampleSheet()
 {
+	if (is_batch_view_)
+	{
+		GUIHelper::showMessage("SampleSheet export", "SampleSheet export is not supported in batch view!");
+		return;
+	}
 	NGSD db;
 	try
 	{
+		//name
 		QString name = ui_->name->text();
 		if (name.startsWith("#")) name.remove(0,1);
+
+		//get settings from user
+		NsxSettingsDialog dlg(this);
+		if (dlg.exec()==QDialog::Rejected) return;
+		NsxAnalysisSettings settings = dlg.getSettings();
+
+		//create sample sheet
 		QString output_path = Settings::string("sample_sheet_path") + "/" + name + ".csv";
 		QStringList warnings;
-		QString sample_sheet = db.createSampleSheet(Helper::toInt(run_id_, "Sequencing run id"), warnings);
+		QString sample_sheet = db.createSampleSheet(Helper::toInt(run_ids_.at(0), "Sequencing run id"), warnings, settings);
 
+		//show warnings
 		if (warnings.size() > 0)
 		{
 			QMessageBox::StandardButton reply;
@@ -507,12 +752,11 @@ void SequencingRunWidget::exportSampleSheet()
 			if (reply== QMessageBox::No) return;
 		}
 
+		//store file
 		QSharedPointer<QFile> output_file = Helper::openFileForWriting(output_path);
 		output_file->write(sample_sheet.toLatin1());
-		output_file->flush();
-		output_file->close();
 
-		QMessageBox::information(this, "SampleSheet exported", "SampleSheet for NovaSeq X Plus exported!");
+		QMessageBox::information(this, "SampleSheet exported", "SampleSheet for NovaSeq X Plus written to:\n" +output_path);
 	}
 	catch (Exception& e)
 	{
@@ -520,7 +764,7 @@ void SequencingRunWidget::exportSampleSheet()
 	}
 }
 
-void SequencingRunWidget::setQCMetricAccessions(const QSet<QString>& sample_types)
+void SequencingRunWidget::setQCMetricAccessions(const QSet<QString>& sample_types, const QSet<QString>& system_types)
 {
 	// determine QC parameter based on sample types
 	qc_metric_accessions_.clear();
@@ -532,6 +776,10 @@ void SequencingRunWidget::setQCMetricAccessions(const QSet<QString>& sample_type
 	if (sample_types.contains("DNA") || sample_types.contains("RNA") || sample_types.contains("DNA (amplicon)") || sample_types.contains("DNA (native)"))
 	{
 		qc_metric_accessions_ << "QC:2000027"; // target region 20x percentage
+	}
+	if (system_types.contains("lrGS"))
+	{
+		qc_metric_accessions_ << "QC:2000131"; // N50 value
 	}
 	if (sample_types.contains("cfDNA"))
 	{
@@ -592,16 +840,19 @@ void SequencingRunWidget::updateReadQualityTable()
 
 	//use different layout depending on sequencer
 	NGSD db;
-	QString device_type = db.getValue("SELECT d.type FROM sequencing_run r, device d WHERE r.device_id=d.id AND r.id=:0", false, run_id_).toString();
+	QString device_type = db.getValue("SELECT d.type FROM sequencing_run r, device d WHERE r.device_id=d.id AND r.id=:0", false, run_ids_.at(0)).toString();
 	if (device_type == "PromethION")
 	{
 		//set headers
 
 		//set horizontal header
-		table->setColumnCount(1);
+		table->setColumnCount(run_ids_.size());
 		table->horizontalHeader()->setVisible(true);
-		table->setHorizontalHeaderItem(0, GUIHelper::createTableItem(db.getValue("SELECT name FROM sequencing_run WHERE id=:0", false, run_id_).toString(), Qt::AlignCenter));
-
+		int c = 0;
+		foreach (const QString& run_id, run_ids_)
+		{
+			table->setHorizontalHeaderItem(c++, GUIHelper::createTableItem(db.getValue("SELECT name FROM sequencing_run WHERE id=:0", false, run_id).toString(), Qt::AlignCenter));
+		}
 
 		//set vertical header
 		QStringList v_headers;
@@ -614,29 +865,37 @@ void SequencingRunWidget::updateReadQualityTable()
 		}
 
 		//get ONT QC
+		c = 0;
 		SqlQuery qc_query = db.getQuery();
-		qc_query.exec("SELECT * FROM runqc_ont WHERE sequencing_run_id=" + run_id_);
+		foreach (const QString& run_id, run_ids_)
+		{
+			qc_query.exec("SELECT * FROM runqc_ont WHERE sequencing_run_id=" + run_id);
 
-		if (qc_query.size() > 1)
-		{
-			THROW(ProgrammingException, "Multiple QC entries found. This should not happen!");
+			if (qc_query.size() > 1)
+			{
+				THROW(ProgrammingException, "Multiple QC entries found. This should not happen!");
+			}
+			else if (qc_query.size() == 1)
+			{
+				qc_query.next();
+				int r = 0;
+				table->setItem(r++, c, GUIHelper::createTableItem(qc_query.value("read_num").toInt()));
+				table->setItem(r++, c, GUIHelper::createTableItem(qc_query.value("yield").toDouble() / 1e9, 2));
+				table->setItem(r++, c, GUIHelper::createTableItem(qc_query.value("passing_filter_perc").toDouble(), 2));
+				table->setItem(r++, c, GUIHelper::createTableItem(qc_query.value("fraction_skipped").toDouble() * 100.0, 2));
+				table->setItem(r++, c, GUIHelper::createTableItem(qc_query.value("q20_perc").toDouble(), 2));
+				table->setItem(r++, c, GUIHelper::createTableItem(qc_query.value("q30_perc").toDouble(), 2));
+				table->setItem(r++, c, GUIHelper::createTableItem(qc_query.value("n50").toInt()));
+				c++;
+			}
 		}
-		else if (qc_query.size() == 1)
-		{
-			qc_query.next();
-			int r = 0;
-			table->setItem(r++, 0, GUIHelper::createTableItem(qc_query.value("read_num").toInt()));
-			table->setItem(r++, 0, GUIHelper::createTableItem(qc_query.value("yield").toDouble() / 1e9, 2));
-			table->setItem(r++, 0, GUIHelper::createTableItem(qc_query.value("passing_filter_perc").toDouble(), 2));
-			table->setItem(r++, 0, GUIHelper::createTableItem(qc_query.value("fraction_skipped").toDouble() * 100.0, 2));
-			table->setItem(r++, 0, GUIHelper::createTableItem(qc_query.value("q20_perc").toDouble(), 2));
-			table->setItem(r++, 0, GUIHelper::createTableItem(qc_query.value("q30_perc").toDouble(), 2));
-			table->setItem(r++, 0, GUIHelper::createTableItem(qc_query.value("n50").toInt()));
-		}
+
+
 
 	}
 	else
 	{
+		if (is_batch_view_) THROW(ArgumentException, "Batch view only supported for PromethION runs!");
 		//set header
 		QStringList headers;
 		headers << "read" << "lane" << "Q30 [%]" << "error rate [%]" << "occupied [%]" << "clusters [K/mm2]" << "clusters PF [%]" << "yield [GB]";
@@ -649,7 +908,7 @@ void SequencingRunWidget::updateReadQualityTable()
 
 		//add QC of reads/lanes
 		SqlQuery q_reads = db.getQuery();
-		q_reads.exec("SELECT * FROM runqc_read WHERE sequencing_run_id=" + run_id_ + " ORDER BY read_num ASC");
+		q_reads.exec("SELECT * FROM runqc_read WHERE sequencing_run_id=" + run_ids_.at(0) + " ORDER BY read_num ASC");
 		QHash<QString, QVector<double>> type2values;
 		while(q_reads.next())
 		{