+++ src/cppNGSD-TEST/NGSD_Test.h
@@ -1465,7 +1465,6 @@ private slots:
 		IS_TRUE(esd_test_input.filtered_by_trio_stringent == esd_db_export.filtered_by_trio_stringent);
 		IS_TRUE(esd_test_input.filtered_by_trio_relaxed == esd_db_export.filtered_by_trio_relaxed);
 
-
 		//change input
 		esd_test_input.dna_rna = "DNA_67890";
 
@@ -2951,7 +2950,6 @@ private slots:
 		COMPARE_FILES("out/somatic_report_tumor_normal_2.xml", TESTDATA("data_out/somatic_report_tumor_normal_2.xml"));
 	}
 
-
 	//Test tumor only RTF report generation
 	void report_tumor_only()
 	{
@@ -3419,7 +3417,7 @@ private slots:
 		db.executeQueriesFromFile(TESTDATA("data_in/NGSD_in5.sql"));
 
 		QStringList warnings;
-		QString sample_sheet = db.createSampleSheet(1, warnings);
+		QString sample_sheet = db.createSampleSheet(1, warnings, NsxAnalysisSettings());
 		S_EQUAL(warnings.at(0), "WARNING: The number of lanes covered by samples (5) and the number of lanes on the flow cell (8) does not match!");
 
 		//write to file
@@ -3432,7 +3430,7 @@ private slots:
 
 		//second run without adapter sequence
 		warnings.clear();
-		sample_sheet = db.createSampleSheet(2, warnings);
+		sample_sheet = db.createSampleSheet(2, warnings, NsxAnalysisSettings());
 		S_EQUAL(warnings.at(0), "WARNING: The number of lanes covered by samples (3) and the number of lanes on the flow cell (2) does not match!");
 		S_EQUAL(warnings.at(1), "WARNING: No adapter for read 1 provided! Adapter trimming will not work.");
 		S_EQUAL(warnings.at(2), "WARNING: No adapter for read 2 provided! Adapter trimming will not work.");
@@ -3446,4 +3444,5 @@ private slots:
 		COMPARE_FILES("out/NovaSeqX_samplesheet2.csv",  TESTDATA("data_out/NovaSeqX_samplesheet2.csv") );
 
 	}
+
 };